restructure cpp headers
